\name{simu3}
\alias{simu3}
\docType{data}
\title{
Simulation data 3
}
\description{
This data set is simulated as described in the Simulation Study 3 of the reference. It is a 100 by 20 data frame, representing 20 full rankings. The columns 1-5 and the columns 6-10 represent two highly correlated ranking groups, respectively.
}
\usage{data("simu3")}
\format{
  A data frame with 100 observations on the following 20 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
    \item{\code{V5}}{a numeric vector}
    \item{\code{V6}}{a numeric vector}
    \item{\code{V7}}{a numeric vector}
    \item{\code{V8}}{a numeric vector}
    \item{\code{V9}}{a numeric vector}
    \item{\code{V10}}{a numeric vector}
    \item{\code{V11}}{a numeric vector}
    \item{\code{V12}}{a numeric vector}
    \item{\code{V13}}{a numeric vector}
    \item{\code{V14}}{a numeric vector}
    \item{\code{V15}}{a numeric vector}
    \item{\code{V16}}{a numeric vector}
    \item{\code{V17}}{a numeric vector}
    \item{\code{V18}}{a numeric vector}
    \item{\code{V19}}{a numeric vector}
    \item{\code{V20}}{a numeric vector}
  }
}
\references{
An extended Mallows model for ranked data aggregation
}
\examples{
data(simu3)
dim(simu3)
}
\keyword{datasets}
